Move the adapter binary to a separate subdir
