+++ .gitignore
@@ -1,4 +1,4 @@
 *.swp
 *~
 vendor
-adapter
+./adapter
