+++ cmd/adapter/app/start.go
